requirements.txt
gunicorn
discord.py
flask
python-dotenv
discord.py[voice]
discord-ext-voice-recv
PyNaCl
